#76 Rename module (translate → deepl)
